+++ TDEngine2/plugins/D3D11GraphicsContext/source/CD3D11Buffer.cpp
@@ -68,7 +68,10 @@ namespace TDEngine2
 		}
 
 #if TDE2_DEBUG_MODE
-		pBuffer->SetPrivateData(WKPDID_D3DDebugObjectName, static_cast<U32>(params.mName.length()), params.mName.c_str());
+		if (params.mName)
+		{
+			pBuffer->SetPrivateData(WKPDID_D3DDebugObjectName, static_cast<U32>(strlen(params.mName)), params.mName);
+		}
 #endif
 
 		return Wrench::TOkValue<ID3D11Buffer*>(pBuffer);