+++ TDEngine2/plugins/D3D11GraphicsContext/source/CD3D11Texture.cpp
@@ -72,7 +72,10 @@ namespace TDEngine2
 		}
 
 #if TDE2_DEBUG_MODE
-		pTexture->SetPrivateData(WKPDID_D3DDebugObjectName, static_cast<U32>(strlen(params.mName)), params.mName);
+		if (params.mName)
+		{
+			pTexture->SetPrivateData(WKPDID_D3DDebugObjectName, static_cast<U32>(strlen(params.mName)), params.mName);
+		}
 #endif
 
 		return Wrench::TOkValue<ID3D11Texture2D*>(pTexture);