+++ TDEngine2/plugins/OGLGraphicsContext/include/COGLGraphicsObjectManager.h
@@ -194,5 +194,8 @@ namespace TDEngine2
 
 			TNativeTexturesArray     mpTexturesArray;
 			TNativeBuffersArray      mpBuffersArray;
+
+			std::unordered_map<U32, std::vector<TTextureHandleId>> mTransientTexturesPool;
+			std::unordered_map<U32, std::vector<TBufferHandleId>>  mTransientBuffersPool;
 	};
 }