+++ TDEngine2/plugins/OGLGraphicsContext/source/COGLGraphicsObjectManager.cpp
@@ -24,6 +24,24 @@ namespace TDEngine2
 	{
 		E_RESULT_CODE result = RC_OK;
 
+		if (E_GRAPHICS_RESOURCE_INIT_FLAGS::TRANSIENT == (params.mFlags & E_GRAPHICS_RESOURCE_INIT_FLAGS::TRANSIENT))
+		{
+			const U32 hash = ComputeStateDescHash(params);
+
+			auto&& it = mTransientBuffersPool.find(hash);
+			if (it != mTransientBuffersPool.cend())
+			{
+				auto& availableTransientBuffers = it->second;
+				if (!availableTransientBuffers.empty())
+				{
+					const TBufferHandleId resourceId = availableTransientBuffers.back();
+					availableTransientBuffers.pop_back();
+
+					return TResult<TBufferHandleId>(resourceId);
+				}
+			}
+		}
+
 		TPtr<IBuffer> pBuffer = TPtr<IBuffer>(CreateOGLBuffer(mpGraphicsContext, params, result));
 		if (!pBuffer || RC_OK != result)
 		{
@@ -48,6 +66,24 @@ namespace TDEngine2
 	TResult<TTextureHandleId> COGLGraphicsObjectManager::CreateTexture(const TInitTextureImplParams& params)
 	{
 		E_RESULT_CODE result = RC_OK;
+
+		if (E_GRAPHICS_RESOURCE_INIT_FLAGS::TRANSIENT == (params.mFlags & E_GRAPHICS_RESOURCE_INIT_FLAGS::TRANSIENT))
+		{
+			const U32 hash = ComputeStateDescHash(params);
+
+			auto&& it = mTransientTexturesPool.find(hash);
+			if (it != mTransientTexturesPool.cend())
+			{
+				auto& availableTransientTextures = it->second;
+				if (!availableTransientTextures.empty())
+				{
+					const TTextureHandleId resourceId = availableTransientTextures.back();
+					availableTransientTextures.pop_back();
+
+					return TResult<TTextureHandleId>(resourceId);
+				}
+			}
+		}
 
 		TPtr<ITextureImpl> pTexture = TPtr<ITextureImpl>(CreateOGLTextureImpl(mpGraphicsContext, params, result));
 		if (!pTexture || RC_OK != result)
@@ -83,6 +119,14 @@ namespace TDEngine2
 			return RC_FAIL;
 		}
 
+		const auto& bufferParams = mpBuffersArray[bufferPlacementIndex]->GetParams();
+
+		if (E_GRAPHICS_RESOURCE_INIT_FLAGS::TRANSIENT == (bufferParams.mFlags & E_GRAPHICS_RESOURCE_INIT_FLAGS::TRANSIENT))
+		{
+			mTransientBuffersPool[ComputeStateDescHash(bufferParams)].push_back(bufferHandle);
+			return RC_OK;
+		}
+
 		mpBuffersArray[bufferPlacementIndex] = nullptr;
 
 		return RC_OK;
@@ -95,13 +139,21 @@ namespace TDEngine2
 			return RC_INVALID_ARGS;
 		}
 
-		const USIZE bufferPlacementIndex = static_cast<USIZE>(textureHandle);
-		if (bufferPlacementIndex >= mpTexturesArray.size())
+		const USIZE texturePlacementIndex = static_cast<USIZE>(textureHandle);
+		if (texturePlacementIndex >= mpTexturesArray.size())
 		{
 			return RC_FAIL;
 		}
 
-		mpTexturesArray[bufferPlacementIndex] = nullptr;
+		const auto& textureParams = mpTexturesArray[texturePlacementIndex]->GetParams();
+
+		if (E_GRAPHICS_RESOURCE_INIT_FLAGS::TRANSIENT == (textureParams.mFlags & E_GRAPHICS_RESOURCE_INIT_FLAGS::TRANSIENT))
+		{
+			mTransientTexturesPool[ComputeStateDescHash(textureParams)].push_back(textureHandle);
+			return RC_OK;
+		}
+
+		mpTexturesArray[texturePlacementIndex] = nullptr;
 
 		return RC_OK;
 	}